atualização geral

+++ sistema-tarefas/src/main/java/com/sistema/sistema_tarefas/entity/Usuario.java
@@ -22,8 +22,8 @@ public class Usuario {
     public Usuario() {}
     
   
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public long getId() { return id; }
+    public void setId(long id) { this.id = (int) id; }
 
     public String getNome() { return nome; }
     public void setNome(String nome) { this.nome = nome; }
